Crate tower_buffer

Buffer requests when the inner service is out of capacity.

Buffering works by spawning a new task that is dedicated to pulling requests out of the buffer and dispatching them to the inner service. By adding a buffer and a dedicated task, the Buffer layer in front of the service can be Clone even if the inner service is not.

Currently, Buffer uses an unbounded buffering strategy, which is not a good thing to put in production situations. However, it illustrates the idea and capabilities around adding buffering to an arbitrary Service.

Structs

Buffer
Adds a buffer in front of an inner service.

ResponseFuture
Future eventually completed with the response to the original request.

SpawnError
Error produced when spawning the worker fails

Worker
Task that handles processing the buffer. This type should not be used directly, instead Buffer requires an Executor that can accept this task.

Enums

Error
Errors produced by Buffer.