Crate tower_buffer

Buffer requests when the inner service is out of capacity.

Buffering works by spawning a new task that is dedicated to pulling requests out of the buffer and dispatching them to the inner service. By adding a buffer and a dedicated task, the Buffer layer in front of the service can be Clone even if the inner service is not.

Modules

error
Error types

future
Future types

Structs

Buffer
Adds a buffer in front of an inner service.

BufferLayer
Buffer requests with a bounded buffer